updated newmain, widgets, imports, and data_log
moved around universal datetime source of truth to prevent dependency loop. slots and signals for heading lock button color changes and changing the closed_loop_dict["head"] value accordingly, when this value is 1 there is now an error thrown in generator.py for the code that makes heading lock actually work. this is good. we will fix it. then heading lock will work and be controlled by this button
-Davis